Suppress an signed/unsigned mismatch warning when building with MSVS
Comparing ERROR_PIPE_BUSY with the DWORD result of GetLastError() may
require a cast to avoid a signed/unsigned comparison warning.

c:\projects\breakpad\src\src\testing\googletest\include\gtest\gtest.h(1520): error C2220: warning treated as error - no 'object' file generated (compiling source file crash_generation_server_test.cc) [C:\projects\breakpad\src\src\client\windows\unittests\client_tests.vcxproj]
         c:\projects\breakpad\src\src\testing\googletest\include\gtest\gtest.h(1548): note: see reference to function template instantiation 'testing::AssertionResult testing::internal::CmpHelperEQ<T1,T2>(const char *,const char *,const T1 &,const T2 &)' being compiled
                 with
                 [
                     T1=long,
                     T2=DWORD
                 ] (compiling source file crash_generation_server_test.cc)
         c:\projects\breakpad\src\src\client\windows\unittests\crash_generation_server_test.cc(112): note: see reference to function template instantiation 'testing::AssertionResult testing::internal::EqHelper<false>::Compare<long,DWORD>(const char *,const char *,const T1 &,const T2 &)' being compiled
                 with
                 [
                     T1=long,
                     T2=DWORD
                 ]
         c:\projects\breakpad\src\src\client\windows\unittests\crash_generation_server_test.cc(112): note: see reference to function template instantiation 'testing::AssertionResult testing::internal::EqHelper<false>::Compare<long,DWORD>(const char *,const char *,const T1 &,const T2 &)' being compiled
                 with
                 [
                     T1=long,
                     T2=DWORD
                 ]
c:\projects\breakpad\src\src\testing\googletest\include\gtest\gtest.h(1520): warning C4389: '==': signed/unsigned mismatch (compiling source file crash_generation_server_test.cc) [C:\projects\breakpad\src\src\client\windows\unittests\client_tests.vcxproj]

+++ src/client/windows/unittests/crash_generation_server_test.cc
@@ -109,7 +109,7 @@ class CrashGenerationServerTest : public ::testing::Test {
                              NULL);
 
     if (pipe == INVALID_HANDLE_VALUE) {
-      ASSERT_EQ(ERROR_PIPE_BUSY, GetLastError());
+      ASSERT_EQ(static_cast<DWORD>(ERROR_PIPE_BUSY), GetLastError());
 
       // Cannot continue retrying if wait on pipe fails.
       ASSERT_TRUE(WaitNamedPipe(kPipeName, 500));
